JBIDE-17156 Palette wizards of HTML category should not have 'Add js/css' checkbox Inheritance of Palette wizard is changed so that super class has only functionalyty common for all widgets, while specific of jQuery and HTML is moved to subclasses.
