+++ plugins/org.jboss.tools.jst.web.ui/src/org/jboss/tools/jst/web/ui/palette/html/jquery/wizard/NewFormWizard.java
@@ -13,13 +13,14 @@ package org.jboss.tools.jst.web.ui.palette.html.jquery.wizard;
 import org.jboss.tools.common.model.ui.editors.dnd.DropWizardMessages;
 import org.jboss.tools.common.model.ui.editors.dnd.IElementGenerator.ElementNode;
 import org.jboss.tools.jst.web.ui.JSTWebUIImages;
+import org.jboss.tools.jst.web.ui.palette.html.wizard.NewHTMLWidgetWizard;
 
 /**
  * 
  * @author Viacheslav Kabanovich
  *
  */
-public class NewFormWizard extends NewJQueryWidgetWizard<NewFormWizardPage> implements JQueryConstants {
+public class NewFormWizard extends NewHTMLWidgetWizard<NewFormWizardPage> implements JQueryConstants {
 	static String prefixId = "form-";
 
 	public NewFormWizard() {
