+++ plugins/org.jboss.tools.jst.web.ui/src/org/jboss/tools/jst/web/ui/palette/html/jquery/wizard/NewJQueryWidgetWizardPage.java
@@ -11,14 +11,19 @@
 package org.jboss.tools.jst.web.ui.palette.html.jquery.wizard;
 
 import java.beans.PropertyChangeEvent;
+import java.beans.PropertyChangeListener;
 
+import org.eclipse.jface.dialogs.DialogSettings;
+import org.eclipse.jface.dialogs.IDialogSettings;
+import org.eclipse.swt.layout.GridData;
+import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
-import org.jboss.tools.common.model.ui.editors.dnd.ValidationException;
+import org.eclipse.swt.widgets.Control;
+import org.jboss.tools.common.ui.widget.editor.CheckBoxFieldEditor;
 import org.jboss.tools.common.ui.widget.editor.IFieldEditor;
 import org.jboss.tools.jst.web.kb.internal.taglib.html.jq.JQueryMobileVersion;
-import org.jboss.tools.jst.web.ui.internal.properties.advanced.LayoutUtil;
-import org.jboss.tools.jst.web.ui.internal.properties.advanced.LayoutUtil.TwoColumns;
-import org.jboss.tools.jst.web.ui.palette.html.wizard.AbstractNewHTMLWidgetWizardPage;
+import org.jboss.tools.jst.web.ui.WebUiPlugin;
+import org.jboss.tools.jst.web.ui.palette.html.wizard.VersionedNewHTMLWidgetWizardPage;
 import org.jboss.tools.jst.web.ui.palette.html.wizard.WizardMessages;
 
 /**
@@ -26,9 +31,7 @@ import org.jboss.tools.jst.web.ui.palette.html.wizard.WizardMessages;
  * @author Viacheslav Kabanovich
  *
  */
-public class NewJQueryWidgetWizardPage extends AbstractNewHTMLWidgetWizardPage implements JQueryConstants {
-	IFieldEditor addID = null;
-	IFieldEditor id = null;
+public class NewJQueryWidgetWizardPage extends VersionedNewHTMLWidgetWizardPage implements JQueryConstants {
 
 	public NewJQueryWidgetWizardPage(String pageName, String title) {
 		super(pageName, title);
@@ -43,63 +46,42 @@ public class NewJQueryWidgetWizardPage extends AbstractNewHTMLWidgetWizardPage i
 		return getWizard().getVersion();
 	}
 
-	protected void createIDEditor(Composite parent, boolean enabling) {
-		if(!enabling) {
-			id = JQueryFieldEditorFactory.createIDEditor();
-			addEditor(id, parent);
+	protected IFieldEditor createAddLibsEditor(Composite parent) {
+		boolean addJSCSS = true; 
+		IDialogSettings settings = WebUiPlugin.getDefault().getDialogSettings();
+		IDialogSettings insertTagSettings = settings.getSection(SECTION_NAME);
+		if(insertTagSettings != null) {
+			addJSCSS = insertTagSettings.getBoolean(ADD_JS_CSS_SETTING_NAME);
 		} else {
-			addID = JQueryFieldEditorFactory.createAddIDEditor();
-			addEditor(addID, parent);
-			id = JQueryFieldEditorFactory.createIDEditor2();
-			addEditor(id, parent);
-			updateIDEnablement();
+			insertTagSettings = DialogSettings.getOrCreateSection(settings, SECTION_NAME);
+			insertTagSettings.put(ADD_JS_CSS_SETTING_NAME, true);
 		}
-	}
-
-	protected boolean isTrue(String editorID) {
-		return TRUE.equals(getEditorValue(editorID));
-	}
-
-	public boolean isIDEnabled() {
-		return id != null && (addID == null || TRUE.equals(addID.getValueAsString()));
-	}
-
-	public void propertyChange(PropertyChangeEvent evt) {
-		if(EDITOR_ID_ADD_ID.equals(evt.getPropertyName())) {
-			updateIDEnablement();
-		}
-		super.propertyChange(evt);
-	}
-
-	public void validate() throws ValidationException {
-		if(isIDEnabled()) {
-			String id = getEditorValue(EDITOR_ID_ID);
-			if(id != null && !getWizard().isIDAvailable(id)) {
-				throw new ValidationException(WizardMessages.errorIDisUsed);
+		final IFieldEditor addLibs = new CheckBoxFieldEditor(ADD_JS_CSS_SETTING_NAME, WizardMessages.addReferencesToJSCSSLabel, Boolean.valueOf(addJSCSS)){
+			@Override
+			public void doFillIntoGrid(Object parent) {
+				Composite c = (Composite) parent;
+				final Control[] controls = (Control[]) getEditorControls(c);
+				Button button = (Button)controls[0];
+				button.setText(WizardMessages.addReferencesToJSCSSLabel);
+				button.setToolTipText(WizardMessages.addReferencesToJSCSSTooltip);
+				GridData d = new GridData(GridData.FILL_HORIZONTAL);
+				d.horizontalSpan = 1;
+				d.minimumWidth = 200;
+				button.setLayoutData(d);
 			}
-		}
-	}
-
-	protected void updateIDEnablement() {
-		if(addID != null && id != null && getLeftPanel() != null) {
-			id.setEnabled(isIDEnabled());
-		}
-	}
-
-	protected void setEnabled(String editorName, boolean value) {
-		IFieldEditor editor = getEditor(editorName);
-		if(editor != null && getLeftPanel() != null) {
-			editor.setEnabled(value);
-		}
-	}
-
-	public static Composite[] createColumns(Composite parent, int k) {
-		if(parent == null) return new Composite[k];
-		return LayoutUtil.createColumns(parent, k);
-	}
-
-	public static TwoColumns createTwoColumns(Composite parent) {
-		return parent == null ? new TwoColumns(null, null) : LayoutUtil.createTwoColumns(parent);
+		};
+		addLibs.doFillIntoGrid(parent);
+		addEditor(addLibs);
+		addLibs.addPropertyChangeListener(new PropertyChangeListener() {
+			@Override
+			public void propertyChange(PropertyChangeEvent evt) {
+				IDialogSettings settings = WebUiPlugin.getDefault().getDialogSettings();
+				IDialogSettings insertTagSettings = settings.getSection(SECTION_NAME);
+				insertTagSettings.put(ADD_JS_CSS_SETTING_NAME, Boolean.parseBoolean(addLibs.getValue().toString()));
+			}
+		});
+		
+		return addLibs;
 	}
 
 }
